requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
python-dotenv
firecrawl-py